pyproject.toml
[project]
name = "deep-research-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "gradio>=5.41.1",
    "openai-agents>=0.2.5",
    "openai[agents]>=1.99.5",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sendgrid>=6.12.4",
]
